assistAIve/requirements.txt
Flask
mysql-connector-python
spacy